requirements.txt
requests==2.32.3
duckdb==1.1.3
apache-airflow[amazon]==2.10.4
fsspec==2024.12.0
s3fs==2024.12.0
minio==7.2.13
pyarrow
python-dotenv==1.0.1

